rewrite to doozer
it took a rewrite of progrium/buildstep to get this to work even remotely like travis expects for Davis-CI. Implemented partial support for .travis.yml and removed the heroku buildpack stuff that didn't make sense.

With that done, the project name didn't make sense anymore, so it's getting renamed Doozer as of this commit. Retaining the file names and structure for now, since Davis-CI still depends on those magic names matching the buildstep api.